- Implemented timestep-method that can be called consistently on all unit types - Removed old gui-classes from the build, as they shall be replaced by a newly designed gui that is consistent with the improved simulation

+++ src/GuiInterface.cpp
@@ -6,7 +6,7 @@ GuiInterface<T, U>::GuiInterface(const pair<double, double> minPos, const pair<d
 {}
 
 template<class T, class U>
-bool GuiInterface<T, U>::run(size_t const intervals)
+bool GuiInterface<T, U>::run(int const intervals)
 {
 	/*
 	while (*intervals == -1)
